+++ sample-apps/react-native/ExpoTikTokApp/components/common/AnnotatedText.tsx
