+++ sample-apps/react-native/ExpoTikTokApp/components/mentions/MentionPreview.tsx
